bconnect_be/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==1.2.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
marshmallow==3.21.1
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
cryptography==42.0.5
python-dotenv==1.0.1
marshmallow-sqlalchemy==1.0.0